move files to basketball_reference_web_scraper_directory
