+++ basketball_reference_web_scraper/client.py
@@ -1,8 +1,8 @@
-import http_client
+from basketball_reference_web_scraper import http_client
 
-from output import box_scores_to_csv, schedule_to_csv
-from output import output
-from json_encoders import BasketballReferenceJSONEncoder
+from basketball_reference_web_scraper.output import box_scores_to_csv, schedule_to_csv
+from basketball_reference_web_scraper.output import output
+from basketball_reference_web_scraper.json_encoders import BasketballReferenceJSONEncoder
 
 
 def box_scores(day, month, year, output_type=None, output_file_path=None, output_write_option=None, json_options=None):
